@@ -6,125 +6,143 @@
     "url": "https://github.com/freecodecamp/freecodecamp.git"
   },
   "scripts": {
-    "build": "gulp build",
+    "only-once": "npm run create-rev && echo '\n\nseeding database\n\n' && node seed && node seed/nonprofits && echo",
+    "create-rev": "node -e \"console.log('\\n\\ncreating manifest\\n\\n'); require('fs').writeFileSync('server/rev-manifest.json', '{}');\"",
+    "build": "NODE_ENV=production gulp build -p",
     "start": "babel-node server/server.js",
-    "prestart-production": "bower cache clean && bower install && gulp build",
+    "prestart-production": "bower cache clean && bower install && gulp build -p",
     "start-production": "node pm2Start",
-    "lint": "eslint --ext=.js,.jsx .",
-    "test": "mocha --compilers js:babel/register"
+    "lint": "npm run lint-js && npm run lint-json",
+    "lint-challenges": "jsonlint -q seed/challenges/**/*.json",
+    "lint-nonprofits": "jsonlint -q seed/nonprofits.json",
+    "lint-server": "jsonlint -q server/*.json",
+    "lint-resources": "jsonlint -q server/resources/*.json",
+    "lint-utils": "jsonlint -q server/utils/*.json",
+    "lint-js": "eslint --ext=.js,.jsx server/ common/ config/ client/",
+    "lint-json": "npm run lint-server && npm run lint-nonprofits && npm run lint-challenges && npm run lint-resources && npm run lint-utils",
+    "test-challenges": "babel-node seed/test-challenges.js | tnyan",
+    "pretest": "npm run lint",
+    "test": "npm run test-challenges"
   },
   "license": "(BSD-3-Clause AND CC-BY-SA-4.0)",
   "dependencies": {
-    "accepts": "~1.2.5",
-    "async": "~0.9.0",
-    "babel": "5.8.23",
-    "babel-core": "5.8.23",
-    "babel-eslint": "4.1.1",
-    "babel-loader": "5.2.2",
-    "bcrypt-nodejs": "~0.0.3",
+    "accepts": "^1.3.0",
+    "async": "^1.5.0",
+    "babel": "5.8.34",
+    "babel-core": "5.8.34",
+    "babel-eslint": "^4.1.4",
+    "babel-loader": "5.4.0",
     "body-parser": "^1.13.2",
-    "chai-jquery": "~2.0.0",
-    "cheerio": "~0.18.0",
+    "chai-jquery": "^2.0.0",
+    "cheerio": "~0.19.0",
     "classnames": "^2.1.2",
-    "clockwork": "~0.1.1",
-    "compression": "~1.2.1",
-    "connect-mongo": "~0.7.0",
-    "cookie-parser": "~1.3.3",
-    "csso": "~1.3.11",
-    "dateformat": "~1.0.11",
-    "debug": "~2.1.0",
-    "dedent": "^0.4.0",
-    "dotenv": "~0.4.0",
-    "errorhandler": "~1.3.0",
-    "es6-map": "^0.1.1",
-    "eslint": "^1.1.0",
-    "eslint-plugin-react": "^3.2.1",
-    "express": "~4.10.4",
+    "compression": "^1.6.0",
+    "connect-mongo": "~0.8.2",
+    "cookie-parser": "^1.4.0",
+    "csso": "^1.4.1",
+    "dateformat": "^1.0.11",
+    "debug": "^2.2.0",
+    "dedent": "~0.6.0",
+    "dotenv": "^1.2.0",
+    "emmet-codemirror": "^1.2.5",
+    "errorhandler": "^1.4.2",
+    "es6-map": "~0.1.1",
+    "eslint": "~1.10.2",
+    "eslint-plugin-react": "^3.7.1",
+    "express": "^4.13.3",
     "express-flash": "~0.0.2",
-    "express-session": "~1.9.2",
+    "express-session": "^1.12.1",
     "express-state": "^1.2.0",
-    "express-validator": "~2.8.0",
-    "fetchr": "^0.5.12",
-    "font-awesome": "~4.3.0",
-    "forever": "~0.14.1",
-    "frameguard": "^0.2.2",
-    "github-api": "~0.7.0",
+    "express-validator": "^2.18.0",
+    "fetchr": "~0.5.12",
+    "forever": "~0.15.1",
+    "frameguard": "~0.2.2",
+    "gulp": "^3.9.0",
+    "gulp-babel": "^5.3.0",
+    "gulp-concat": "^2.6.0",
+    "gulp-eslint": "^1.1.0",
+    "gulp-inject": "^3.0.0",
+    "gulp-jsonlint": "^1.1.0",
     "gulp-less": "^3.0.3",
-    "gulp-minify-css": "~0.5.1",
+    "gulp-minify-css": "^1.2.1",
+    "gulp-nodemon": "^2.0.3",
+    "gulp-notify": "^2.2.0",
+    "gulp-order": "^1.1.1",
+    "gulp-plumber": "^1.0.1",
     "gulp-reduce-file": "0.0.1",
     "gulp-rev": "^6.0.1",
-    "gulp-rev-replace": "^0.4.2",
+    "gulp-rev-replace": "~0.4.2",
+    "gulp-uglify": "^1.5.1",
     "gulp-util": "^3.0.6",
     "gulp-webpack": "^1.5.0",
-    "helmet": "~0.9.0",
-    "helmet-csp": "^0.2.3",
-    "jade": "~1.8.0",
-    "json-loader": "^0.5.2",
-    "less": "~1.7.5",
-    "less-middleware": "~2.0.1",
+    "helmet": "~0.15.0",
+    "helmet-csp": "~0.3.0",
+    "history": "^1.17.0",
+    "jade": "^1.11.0",
+    "json-loader": "~0.5.2",
+    "less": "^2.5.1",
     "lodash": "^3.9.3",
-    "loopback": "https://github.com/FreeCodeCamp/loopback.git#fix/no-password",
-    "loopback-boot": "2.8.2",
-    "loopback-component-passport": "https://github.com/FreeCodeCamp/loopback-component-passport.git#feature/flashfailure",
-    "loopback-connector-mongodb": "^1.10.0",
-    "lusca": "~1.0.2",
-    "method-override": "~2.3.0",
-    "moment": "~2.10.2",
+    "loopback": "^2.22.0",
+    "loopback-boot": "^2.13.0",
+    "loopback-component-passport": "^1.6.0",
+    "loopback-connector-mongodb": "1.13.0",
+    "merge-stream": "^1.0.0",
+    "method-override": "^2.3.0",
+    "moment": "^2.10.2",
     "mongodb": "^2.0.33",
-    "morgan": "~1.5.0",
-    "node-libs-browser": "^0.5.2",
-    "node-slack": "0.0.7",
+    "morgan": "^1.6.1",
     "node-uuid": "^1.4.3",
-    "nodemailer": "~1.3.0",
-    "object.assign": "^3.0.0",
+    "nodemailer": "^1.9.0",
+    "normalize-url": "^1.3.1",
+    "object.assign": "^4.0.3",
     "passport-facebook": "^2.0.0",
-    "passport-github": "^0.1.5",
-    "passport-google-oauth2": "^0.1.6",
+    "passport-github": "^1.0.0",
+    "passport-google-oauth2": "~0.1.6",
     "passport-linkedin-oauth2": "^1.2.1",
     "passport-local": "^1.0.0",
     "passport-oauth": "^1.0.0",
     "passport-twitter": "^1.0.3",
-    "pmx": "^0.3.16",
-    "ramda": "~0.10.0",
-    "react": "^0.13.3",
-    "react-bootstrap": "~0.23.7",
-    "react-motion": "~0.1.0",
-    "react-router": "https://github.com/BerkeleyTrue/react-router#freecodecamp",
-    "react-vimeo": "^0.0.3",
-    "request": "~2.53.0",
+    "pmx": "~0.5.5",
+    "ramda": "~0.18.0",
+    "react": "~0.14.3",
+    "react-bootstrap": "~0.28.1",
+    "react-dom": "~0.14.3",
+    "react-motion": "~0.3.1",
+    "react-router": "^1.0.0",
+    "react-router-bootstrap": "https://github.com/FreeCodeCamp/react-router-bootstrap.git#freecodecamp",
+    "react-vimeo": "~0.0.3",
+    "request": "^2.65.0",
     "rev-del": "^1.0.5",
-    "rx": "^2.5.3",
-    "sanitize-html": "~1.6.1",
+    "rx": "^4.0.0",
+    "sanitize-html": "^1.11.1",
     "sort-keys": "^1.1.1",
-    "source-map-support": "^0.3.2",
-    "thundercats": "^2.1.0",
-    "thundercats-react": "^0.1.0",
-    "twit": "~1.1.20",
-    "uglify-js": "~2.4.15",
-    "validator": "~3.22.1",
+    "store": "https://github.com/berkeleytrue/store.js.git#feature/noop-server",
+    "thundercats": "^3.0.0",
+    "thundercats-react": "~0.4.0",
+    "twit": "^2.1.1",
+    "uglify-js": "^2.5.0",
+    "url-regex": "^3.0.0",
+    "validator": "^4.2.1",
     "webpack": "^1.9.12",
-    "yui": "~3.18.1"
+    "webpack-stream": "^3.1.0",
+    "xss-filters": "^1.2.6",
+    "yargs": "^3.30.0"
   },
   "devDependencies": {
-    "blessed": "~0.0.37",
+    "blessed": "~0.1.81",
     "bower-main-files": "~0.0.4",
-    "browser-sync": "~1.8.1",
-    "browserify": "^10.2.4",
-    "chai": "~1.10.0",
+    "browser-sync": "^2.9.12",
+    "chai": "^3.4.0",
     "envify": "^3.4.0",
-    "gulp": "~3.8.8",
-    "gulp-eslint": "~0.9.0",
-    "gulp-inject": "~1.0.2",
-    "gulp-jsonlint": "^1.1.0",
-    "gulp-nodemon": "^2.0.3",
-    "gulp-notify": "^2.2.0",
-    "gulp-plumber": "^1.0.1",
-    "istanbul": "^0.3.15",
-    "loopback-explorer": "^1.7.2",
+    "gulp-sourcemaps": "^1.6.0",
+    "gulp-tape": "0.0.7",
+    "istanbul": "~0.4.0",
+    "jsonlint": "^1.6.2",
+    "loopback-component-explorer": "^2.1.1",
     "loopback-testing": "^1.1.0",
-    "mocha": "~2.0.1",
-    "multiline": "~1.0.1",
-    "supertest": "~0.15.0",
-    "vinyl-source-stream": "^1.1.0"
+    "mocha": "^2.3.3",
+    "tap-nyan": "0.0.2",
+    "tap-spec": "^4.1.1",
+    "tape": "^4.2.2"
   }
 }
