@@ -4,7 +4,7 @@ import React from 'react';
 import Fetchr from 'fetchr';
 import debugFactory from 'debug';
 import { Router } from 'react-router';
-import { history } from 'react-router/lib/BrowserHistory';
+import { createLocation, createHistory } from 'history';
 import { hydrate } from 'thundercats';
 import { Render } from 'thundercats-react';
 
@@ -18,21 +18,63 @@ const services = new Fetchr({
 });
 
 Rx.config.longStackSupport = !!debug.enabled;
+const history = createHistory();
+const appLocation = createLocation(
+  location.pathname + location.search
+);
+
+function location$(history) {
+  return Rx.Observable.create(function(observer) {
+    const dispose = history.listen(function(location) {
+      observer.onNext(location.pathname);
+    });
+
+    return Rx.Disposable.create(() => {
+      dispose();
+    });
+  });
+}
 
 // returns an observable
-app$(history)
+app$({ history, location: appLocation })
   .flatMap(
     ({ AppCat }) => {
+      // instantiate the cat with service
       const appCat = AppCat(null, services);
+      // hydrate the stores
       return hydrate(appCat, catState)
         .map(() => appCat);
     },
-    ({ initialState }, appCat) => ({ initialState, appCat })
+    // not using nextLocation at the moment but will be used for
+    // redirects in the future
+    ({ nextLocation, props }, appCat) => ({ nextLocation, props, appCat })
   )
-  .flatMap(({ initialState, appCat }) => {
+  .doOnNext(({ appCat }) => {
+    const appActions = appCat.getActions('appActions');
+
+    location$(history)
+      .pluck('pathname')
+      .distinctUntilChanged()
+      .doOnNext(route => debug('route change', route))
+      .subscribe(route => appActions.updateRoute(route));
+
+    appActions.goBack.subscribe(function() {
+      history.goBack();
+    });
+
+    appActions
+      .updateRoute
+      .pluck('route')
+      .doOnNext(route => debug('update route', route))
+      .subscribe(function(route) {
+        history.pushState(null, route);
+      });
+  })
+  .flatMap(({ props, appCat }) => {
+    props.history = history;
     return Render(
       appCat,
-      React.createElement(Router, initialState),
+      React.createElement(Router, props),
       DOMContianer
     );
   })
@@ -41,7 +83,7 @@ app$(history)
       debug('react rendered');
     },
     err => {
-      debug('an error has occured', err.stack);
+      throw err;
     },
     () => {
       debug('react closed subscription');
