@@ -1,17 +1,17 @@
 import Rx from 'rx';
-import { Router } from 'react-router';
+import { match } from 'react-router';
 import App from './App.jsx';
 import AppCat from './Cat';
 
 import childRoutes from './routes';
 
-const router$ = Rx.Observable.fromNodeCallback(Router.run, Router);
+const route$ = Rx.Observable.fromNodeCallback(match);
 
 const routes = Object.assign({ components: App }, childRoutes);
 
-export default function app$(location) {
-  return router$(routes, location)
-    .map(([initialState, transistion]) => {
-      return { initialState, transistion, AppCat };
+export default function app$({ location, history }) {
+  return route$({ routes, location, history })
+    .map(([nextLocation, props]) => {
+      return { nextLocation, props, AppCat };
     });
 }
