@@ -8,13 +8,17 @@ const initValue = {
   points: 0
 };
 
-export default Store(initValue)
-  .refs({ displayName: 'AppStore' })
-  .init(({ instance: appStore, args: [cat] }) => {
-    const { setUser, setTitle } = cat.getActions('appActions');
+export default Store({
+  refs: {
+    displayName: 'AppStore',
+    value: initValue
+  },
+  init({ instance: appStore, args: [cat] }) {
+    const { updateRoute, setUser, setTitle } = cat.getActions('appActions');
     const register = createRegistrar(appStore);
 
-    register(setter(fromMany(setUser, setTitle)));
+    register(setter(fromMany(setUser, setTitle, updateRoute)));
 
     return appStore;
-  });
+  }
+});
