@@ -5,12 +5,16 @@ const initialValue = {
   currentHike: {}
 };
 
-export default Store(initialValue)
-  .refs({ displayName: 'HikesStore'})
-  .init(({ instance: hikeStore, args: [cat] }) => {
+export default Store({
+  refs: {
+    displayName: 'HikesStore',
+    value: initialValue
+  },
+  init({ instance: hikeStore, args: [cat] }) {
 
     let { setHikes } = cat.getActions('hikesActions');
     hikeStore.register(setHikes);
 
     return hikeStore;
-  });
+  }
+});
