@@ -6,12 +6,37 @@ const {
   transformer
 } = Store;
 
-export default Store()
-  .refs({ displayName: 'JobsStore' })
-  .init(({ instance: jobsStore, args: [cat] }) => {
-    const { setJobs, findJob, setError } = cat.getActions('JobActions');
+export default Store({
+  refs: {
+    displayName: 'JobsStore',
+    value: { showModal: false }
+  },
+  init({ instance: jobsStore, args: [cat] }) {
+    const {
+      setJobs,
+      findJob,
+      setError,
+      openModal,
+      closeModal,
+      handleForm,
+      setForm,
+      setFollowersCount,
+      setPromoCode,
+      applyPromo,
+      clearPromo
+    } = cat.getActions('JobActions');
     const register = createRegistrar(jobsStore);
     register(setter(setJobs));
-    register(transformer(findJob));
     register(setter(setError));
-  });
+    register(setter(openModal));
+    register(setter(closeModal));
+    register(setter(setForm));
+    register(setter(setPromoCode));
+    register(setter(applyPromo));
+    register(setter(clearPromo));
+    register(setter(setFollowersCount));
+
+    register(transformer(findJob));
+    register(handleForm);
+  }
+});
