@@ -4,52 +4,114 @@ import moment from 'moment';
 import { Observable, Scheduler } from 'rx';
 import assign from 'object.assign';
 import debugFactory from 'debug';
-import utils from '../utils';
+
+import {
+  dasherize,
+  unDasherize,
+  getMDNLinks,
+  randomVerb,
+  randomPhrase,
+  randomCompliment
+} from '../utils';
 
 import {
   saveUser,
   observeMethod,
-  observableQueryFromModel
+  observeQuery
 } from '../utils/rx';
 
 import {
-  userMigration,
   ifNoUserSend
 } from '../utils/middleware';
 
+import getFromDisk$ from '../utils/getFromDisk$';
+
+const isDev = process.env.NODE_ENV !== 'production';
+const isBeta = !!process.env.BETA;
 const debug = debugFactory('freecc:challenges');
 const challengesRegex = /^(bonfire|waypoint|zipline|basejump)/i;
-const firstChallenge = 'waypoint-say-hello-to-html-elements';
+const firstChallenge = 'waypoint-learn-how-free-code-camp-works';
 const challengeView = {
   0: 'coursewares/showHTML',
   1: 'coursewares/showJS',
   2: 'coursewares/showVideo',
   3: 'coursewares/showZiplineOrBasejump',
   4: 'coursewares/showZiplineOrBasejump',
-  5: 'coursewares/showBonfire'
+  5: 'coursewares/showBonfire',
+  7: 'coursewares/showStep'
 };
 
-const dasherize = utils.dasherize;
-const unDasherize = utils.unDasherize;
-const getMDNLinks = utils.getMDNLinks;
+function isChallengeCompleted(user, challengeId) {
+  if (!user) {
+    return false;
+  }
+  return user.completedChallenges.some(challenge =>
+    challenge.id === challengeId );
+}
 
 function numberWithCommas(x) {
   return x.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
 }
 
 function updateUserProgress(user, challengeId, completedChallenge) {
-  const alreadyCompleted = user.completedChallenges.some(({ id }) => {
-    return id === challengeId;
+  let { completedChallenges } = user;
+
+  const indexOfChallenge = _.findIndex(completedChallenges, {
+    id: challengeId
   });
+
+  const alreadyCompleted = indexOfChallenge !== -1;
 
   if (!alreadyCompleted) {
     user.progressTimestamps.push({
       timestamp: Date.now(),
-      completedChallenge
+      completedChallenge: challengeId
     });
+    user.completedChallenges.push(completedChallenge);
+    return user;
   }
-  user.completedChallenges.push(completedChallenge);
+
+  const oldCompletedChallenge = completedChallenges[indexOfChallenge];
+  user.completedChallenges[indexOfChallenge] =
+    Object.assign(
+      {},
+      completedChallenge,
+      {
+        completedDate: oldCompletedChallenge.completedDate,
+        lastUpdated: completedChallenge.completedDate
+      }
+    );
   return user;
+}
+
+
+// small helper function to determine whether to mark something as new
+const dateFormat = 'MMM MMMM DD, YYYY';
+function shouldShowNew(element, block) {
+  if (element) {
+    return typeof element.releasedOn !== 'undefined' &&
+      moment(element.releasedOn, dateFormat).diff(moment(), 'days') >= -30;
+  }
+
+  if (block) {
+    const newCount = block.reduce((sum, { markNew }) => {
+      if (markNew) {
+        return sum + 1;
+      }
+      return sum;
+    }, 0);
+    return newCount / block.length * 100 === 100;
+  }
+}
+
+// meant to be used with a filter method
+// on an array or observable stream
+// true if challenge should be passed through
+// false if should filter challenge out of array or stream
+function shouldNotFilterComingSoon({ isComingSoon, isBeta: challengeIsBeta }) {
+  return isDev ||
+    !isComingSoon ||
+    (isBeta && challengeIsBeta);
 }
 
 module.exports = function(app) {
@@ -57,6 +119,7 @@ module.exports = function(app) {
 
   const challengesQuery = {
     order: [
+      'superOrder ASC',
       'order ASC',
       'suborder ASC'
     ]
@@ -68,30 +131,36 @@ module.exports = function(app) {
   const findChallenge$ = observeMethod(Challenge, 'find');
   // create a stream of all the challenges
   const challenge$ = findChallenge$(challengesQuery)
-    .doOnNext(() => debug('query challenges'))
     .flatMap(challenges => Observable.from(
       challenges,
       null,
       null,
       Scheduler.default
     ))
+    // filter out all challenges that have isBeta flag set
+    // except in development or beta site
+    .filter(challenge => isDev || isBeta || !challenge.isBeta)
     .shareReplay();
 
   // create a stream of challenge blocks
   const blocks$ = challenge$
     .map(challenge => challenge.toJSON())
+    .filter(shouldNotFilterComingSoon)
     // group challenges by block | returns a stream of observables
     .groupBy(challenge => challenge.block)
     // turn block group stream into an array
-    .flatMap(block$ => block$.toArray())
+    .flatMap(blocks$ => blocks$.toArray())
     // turn array into stream of object
-    .map(blockArray => ({
-      name: blockArray[0].block,
-      dashedName: dasherize(blockArray[0].block),
-      challenges: blockArray
+    .map(blocksArray => ({
+      name: blocksArray[0].block,
+      dashedName: dasherize(blocksArray[0].block),
+      challenges: blocksArray,
+      superBlock: blocksArray[0].superBlock,
+      order: blocksArray[0].order
     }))
-    .filter(({ name })=> {
-      return name !== 'Hikes';
+    // filter out hikes
+    .filter(({ superBlock }) => {
+      return !(/hikes/gi).test(superBlock);
     })
     .shareReplay();
 
@@ -110,14 +179,7 @@ module.exports = function(app) {
     send200toNonUser,
     completedZiplineOrBasejump
   );
-  router.post(
-    '/completed-bonfire',
-    send200toNonUser,
-    completedBonfire
-  );
 
-  // the follow routes are covered by userMigration
-  router.use(userMigration);
   router.get('/map', challengeMap);
   router.get(
     '/challenges/next-challenge',
@@ -136,13 +198,17 @@ module.exports = function(app) {
     // find challenge
     return challenge$
       .map(challenge => challenge.toJSON())
-      .filter(({ block }) => block !== 'Hikes')
+      // filter out challenges coming soon
+      .filter(shouldNotFilterComingSoon)
+      // filter out hikes
+      .filter(({ superBlock }) => !(/hikes/gi).test(superBlock))
       .filter(({ id }) => id === challengeId)
       // now lets find the block it belongs to
       .flatMap(challenge => {
         // find the index of the block this challenge resides in
         const blockIndex$ = blocks$
           .findIndex(({ name }) => name === challenge.block);
+
 
         return blockIndex$
           .flatMap(blockIndex => {
@@ -152,11 +218,12 @@ module.exports = function(app) {
                 'could not find challenge block for ' + challenge.block
               );
             }
-            const nextBlock$ = blocks$.elementAt(blockIndex + 1);
-            const firstChallengeOfNextBlock$ = nextBlock$
-              .map(block => block.challenges[0]);
+            const firstChallengeOfNextBlock$ = blocks$
+              .elementAt(blockIndex + 1, {})
+              .map(({ challenges = [] }) => challenges[0]);
 
             return blocks$
+              .filter(shouldNotFilterComingSoon)
               .elementAt(blockIndex)
               .flatMap(block => {
                 // find where our challenge lies in the block
@@ -183,6 +250,9 @@ module.exports = function(app) {
           });
       })
       .map(nextChallenge => {
+        if (!nextChallenge) {
+          return null;
+        }
         nextChallengeName = nextChallenge.dashedName;
         return nextChallengeName;
       })
@@ -195,8 +265,8 @@ module.exports = function(app) {
             req.flash('info', {
               msg: dedent`
                 Once you have completed all of our challenges, you should
-                join our <a href=\"//gitter.im/freecodecamp/HalfWayClub\"
-                target=\"_blank\">Half Way Club</a> and start getting
+                join our <a href="https://gitter.im/freecodecamp/HalfWayClub"
+                target="_blank">Half Way Club</a> and start getting
                 ready for our nonprofit projects.
               `.split('\n').join(' ')
             });
@@ -209,6 +279,7 @@ module.exports = function(app) {
 
   function returnIndividualChallenge(req, res, next) {
     const origChallengeName = req.params.challengeName;
+    const solutionCode = req.query.solution;
     const unDashedName = unDasherize(origChallengeName);
 
     const challengeName = challengesRegex.test(unDashedName) ?
@@ -220,9 +291,16 @@ module.exports = function(app) {
     debug('looking for %s', testChallengeName);
     challenge$
       .filter((challenge) => {
-        return testChallengeName.test(challenge.name);
+        return testChallengeName.test(challenge.name) &&
+          shouldNotFilterComingSoon(challenge);
       })
-      .lastOrDefault(null)
+      .last({ defaultValue: null })
+      .flatMap(challenge => {
+        if (challenge && isDev) {
+          return getFromDisk$(challenge);
+        }
+        return Observable.just(challenge);
+      })
       .flatMap(challenge => {
 
         // Handle not found
@@ -234,35 +312,50 @@ module.exports = function(app) {
               origChallengeName +
               '` Please double check the name.'
           });
-          return Observable.just('/challenges');
+          return Observable.just('/map');
         }
 
         if (dasherize(challenge.name) !== origChallengeName) {
-          return Observable.just('/challenges/' + dasherize(challenge.name));
+          let redirectUrl = `/challenges/${dasherize(challenge.name)}`;
+
+          if (solutionCode) {
+            redirectUrl += `?solution=${encodeURIComponent(solutionCode)}`;
+          }
+
+          return Observable.just(redirectUrl);
         }
 
         // save user does nothing if user does not exist
         return Observable.just({
+
           title: challenge.name,
-          dashedName: origChallengeName,
           name: challenge.name,
           details: challenge.description,
-          tests: challenge.tests,
-          challengeSeed: challenge.challengeSeed,
-          verb: utils.randomVerb(),
-          phrase: utils.randomPhrase(),
-          compliment: utils.randomCompliment(),
+          description: challenge.description,
           challengeId: challenge.id,
           challengeType: challenge.challengeType,
+          dashedName: origChallengeName,
+
+          challengeSeed: challenge.challengeSeed,
+          head: challenge.head,
+          tail: challenge.tail,
+          tests: challenge.tests,
+
+          // identifies if a challenge is completed
+          isCompleted: isChallengeCompleted(req.user, challenge.id),
+
           // video challenges
           video: challenge.challengeSeed[0],
+
           // bonfires specific
-          difficulty: Math.floor(+challenge.difficulty),
           bonfires: challenge,
           MDNkeys: challenge.MDNlinks,
           MDNlinks: getMDNLinks(challenge.MDNlinks),
+
           // htmls specific
-          environment: utils.whichEnvironment()
+          verb: randomVerb(),
+          phrase: randomPhrase(),
+          compliment: randomCompliment()
         });
       })
       .subscribe(
@@ -279,95 +372,25 @@ module.exports = function(app) {
       );
   }
 
-  function completedBonfire(req, res, next) {
-    debug('compltedBonfire');
-    var completedWith = req.body.challengeInfo.completedWith || false;
-    var challengeId = req.body.challengeInfo.challengeId;
-
-    var challengeData = {
-      id: challengeId,
-      name: req.body.challengeInfo.challengeName || '',
-      completedDate: Math.round(+new Date()),
-      solution: req.body.challengeInfo.solution,
-      challengeType: 5
-    };
-
-    observableQueryFromModel(
-        User,
-        'findOne',
-        { where: { username: ('' + completedWith).toLowerCase() } }
-      )
-      .doOnNext(function(pairedWith) {
-        debug('paired with ', pairedWith);
-        if (pairedWith) {
-          updateUserProgress(
-            pairedWith,
-            challengeId,
-            assign({ completedWith: req.user.id }, challengeData)
-          );
-        }
-      })
-      .withLatestFrom(
-        Observable.just(req.user),
-        function(pairedWith, user) {
-          return {
-            user: user,
-            pairedWith: pairedWith
-          };
-        }
-      )
-      // side effects should always be done in do's and taps
-      .doOnNext(function(dats) {
-        updateUserProgress(
-          dats.user,
-          challengeId,
-          dats.pairedWith ?
-            // paired programmer found and adding to data
-            assign({ completedWith: dats.pairedWith.id }, challengeData) :
-            // user said they paired, but pair wasn't found
-            challengeData
-        );
-      })
-      // iterate users
-      .flatMap(function(dats) {
-        debug('flatmap');
-        return Observable.from([dats.user, dats.pairedWith]);
-      })
-      // save user
-      .flatMap(function(user) {
-        // save user will do nothing if user is falsey
-        return saveUser(user);
-      })
-      .subscribe(
-        function(user) {
-          debug('onNext');
-          if (user) {
-            debug('user %s saved', user.username);
-          }
-        },
-        next,
-        function() {
-          debug('completed');
-          return res.status(200).send(true);
-        }
-      );
-  }
-
   function completedChallenge(req, res, next) {
 
     const completedDate = Math.round(+new Date());
-    const { id, name } = req.body;
-    const { challengeId, challengeName } = req.body.challengeInfo || {};
+    const {
+      id,
+      name,
+      challengeType,
+      solution
+    } = req.body;
 
     updateUserProgress(
       req.user,
-      id || challengeId,
+      id,
       {
-        id: id || challengeId,
-        completedDate: completedDate,
-        name: name || challengeName || '',
-        solution: null,
-        githubLink: null,
+        id,
+        challengeType,
+        solution,
+        name,
+        completedDate,
         verified: true
       }
     );
@@ -420,7 +443,7 @@ module.exports = function(app) {
       verified: false
     };
 
-    observableQueryFromModel(
+    observeQuery(
         User,
         'findOne',
         { where: { username: completedWith.toLowerCase() } }
@@ -471,6 +494,8 @@ module.exports = function(app) {
   }
 
   function challengeMap({ user = {} }, res, next) {
+
+    let lastCompleted;
     const daysRunning = moment().diff(new Date('10/15/2014'), 'days');
 
     // if user
@@ -483,13 +508,14 @@ module.exports = function(app) {
       .map(camperCount => numberWithCommas(camperCount));
 
     // create a stream of an array of all the challenge blocks
-    const blocks$ = challenge$
+    const superBlocks$ = challenge$
       // mark challenge completed
       .map(challengeModel => {
         const challenge = challengeModel.toJSON();
         if (completedChallenges.indexOf(challenge.id) !== -1) {
           challenge.completed = true;
         }
+        challenge.markNew = shouldShowNew(challenge);
         return challenge;
       })
       // group challenges by block | returns a stream of observables
@@ -503,30 +529,58 @@ module.exports = function(app) {
           }
           return sum;
         }, 0);
+        const isBeta = _.every(blockArray, 'isBeta');
+        const isComingSoon = _.every(blockArray, 'isComingSoon');
 
         return {
+          isBeta,
+          isComingSoon,
           name: blockArray[0].block,
+          superBlock: blockArray[0].superBlock,
           dashedName: dasherize(blockArray[0].block),
+          markNew: shouldShowNew(null, blockArray),
           challenges: blockArray,
-          completed: completedCount / blockArray.length * 100
+          completed: completedCount / blockArray.length * 100,
+          time: blockArray[0] && blockArray[0].time || '???'
         };
       })
-      .filter(({ name }) => name !== 'Hikes')
+      // filter out hikes
+      .filter(({ superBlock }) => {
+        return !(/hikes/i).test(superBlock);
+      })
       // turn stream of blocks into a stream of an array
+      .toArray()
+      .doOnNext(blocks => {
+        const lastCompletedBlock = _.findLast(blocks, (block) => {
+          return block.completed === 100;
+        });
+        lastCompleted = lastCompletedBlock && lastCompletedBlock.name || null;
+      })
+      .flatMap(blocks => Observable.from(blocks, null, null, Scheduler.default))
+      .groupBy(block => block.superBlock)
+      .flatMap(blocks$ => blocks$.toArray())
+      .map(superBlockArray => ({
+        name: superBlockArray[0].superBlock,
+        blocks: superBlockArray
+      }))
       .toArray();
 
     Observable.combineLatest(
       camperCount$,
-      blocks$,
-      (camperCount, blocks) => ({ camperCount, blocks })
+      superBlocks$,
+      (camperCount, superBlocks) => ({ camperCount, superBlocks })
     )
       .subscribe(
-        ({ camperCount, blocks }) => {
+        ({ camperCount, superBlocks }) => {
           res.render('challengeMap/show', {
-            blocks,
+            superBlocks,
             daysRunning,
+            globalCompletedCount: numberWithCommas(
+              5612952 + (Math.floor((Date.now() - 1446268581061) / 2000))
+            ),
             camperCount,
-            title: "A map of all Free Code Camp's Challenges"
+            lastCompleted,
+            title: 'A Map to Learn to Code and Become a Software Engineer'
           });
         },
         next
