@@ -4,7 +4,8 @@ module.exports = function(app) {
   router.get('/nonprofit-project-instructions', function(req, res) {
     res.redirect(
       301,
-      "//github.com/FreeCodeCamp/freecodecamp/wiki/How-Free-Code-Camp's-Nonprofit-Projects-work"
+      '//github.com/FreeCodeCamp/freecodecamp/wiki/' +
+        "How-Free-Code-Camp's-Nonprofit-Projects-work"
     );
   });
 
@@ -12,9 +13,15 @@ module.exports = function(app) {
     res.redirect(301, '/pmi-acp-agile-project-managers');
   });
 
+  router.get('/wiki', function(req, res) {
+    res.redirect(301, '//github.com/freecodecamp/freecodecamp/wiki');
+  });
+
   router.get('/privacy', function(req, res) {
     res.redirect(
-      301, "//github.com/FreeCodeCamp/freecodecamp/wiki/Free-Code-Camp's-Privacy-Policy"
+      301,
+      '//github.com/FreeCodeCamp/freecodecamp/wiki/' +
+        "Free-Code-Camp's-Privacy-Policy"
     );
   });
 
@@ -23,7 +30,7 @@ module.exports = function(app) {
   });
 
   router.get('/field-guide/*', function(req, res) {
-    res.redirect(302, '//github.com/freecodecamp/freecodecamp/wiki')
+    res.redirect(302, '//github.com/freecodecamp/freecodecamp/wiki');
   });
 
   router.get('/about', function(req, res) {
