@@ -2,23 +2,25 @@ var Rx = require('rx'),
     assign = require('object.assign'),
     sanitizeHtml = require('sanitize-html'),
     moment = require('moment'),
-    mongodb = require('mongodb'),
     debug = require('debug')('freecc:cntr:story'),
     utils = require('../utils'),
     observeMethod = require('../utils/rx').observeMethod,
     saveUser = require('../utils/rx').saveUser,
     saveInstance = require('../utils/rx').saveInstance,
-    MongoClient = mongodb.MongoClient,
-    validator = require('validator'),
-    secrets = require('../../config/secrets');
+    validator = require('validator');
 
-var foundationDate = 1413298800000;
-var time48Hours = 172800000;
+import {
+  ifNoUser401,
+  ifNoUserRedirectTo
+} from '../utils/middleware';
 
-var unDasherize = utils.unDasherize;
-var dasherize = utils.dasherize;
-var getURLTitle = utils.getURLTitle;
-var ifNoUser401 = require('../utils/middleware').ifNoUser401;
+const foundationDate = 1413298800000;
+const time48Hours = 172800000;
+
+const unDasherize = utils.unDasherize;
+const dasherize = utils.dasherize;
+const getURLTitle = utils.getURLTitle;
+const sendNonUserToNews = ifNoUserRedirectTo('/news');
 
 function hotRank(timeValue, rank) {
   /*
@@ -60,10 +62,20 @@ module.exports = function(app) {
   var findStoryById = observeMethod(Story, 'findById');
   var countStories = observeMethod(Story, 'count');
 
+  router.post('/news/userstories', userStories);
   router.get('/news/hot', hotJSON);
+  router.get('/news/feed', RSSFeed);
   router.get('/stories/hotStories', hotJSON);
-  router.get('/stories/submit', submitNew);
-  router.get('/stories/submit/new-story', preSubmit);
+  router.get(
+    '/stories/submit',
+    sendNonUserToNews,
+    submitNew
+  );
+  router.get(
+    '/stories/submit/new-story',
+    sendNonUserToNews,
+    preSubmit
+  );
   router.post('/stories/preliminary', ifNoUser401, newStory);
   router.post('/stories/', ifNoUser401, storySubmission);
   router.get('/news/', hot);
@@ -94,25 +106,54 @@ module.exports = function(app) {
     );
   }
 
+  function RSSFeed(req, res, next) {
+    var query = {
+      order: 'timePosted DESC',
+      limit: 1000
+    };
+    findStory(query).subscribe(
+      function(stories) {
+        var sliceVal = stories.length >= 100 ? 100 : stories.length;
+        var data = stories.sort(sortByRank).slice(0, sliceVal);
+        res.set('Content-Type', 'text/xml');
+        res.render('feed', {
+          title: 'FreeCodeCamp Camper News RSS Feed',
+          description: 'RSS Feed for FreeCodeCamp Top 100 Hot Camper News',
+          url: 'http://www.freecodecamp.com/news',
+          FeedPosts: data
+        });
+      },
+      next
+    );
+  }
+
   function hot(req, res) {
     return res.render('stories/index', {
-      title: 'Hot stories currently trending on Camper News',
+      title: 'Top Stories on Camper News',
       page: 'hot'
     });
   }
 
   function submitNew(req, res) {
+    if (!req.user.isGithubCool) {
+      req.flash('errors', {
+        msg: 'You must link GitHub with your account before you can post' +
+          ' on Camper News.'
+      });
+      return res.redirect('/news');
+    }
+
     return res.render('stories/index', {
       title: 'Submit a new story to Camper News',
       page: 'submit'
     });
   }
 
-  function preSubmit(req, res, next) {
+  function preSubmit(req, res) {
     var data = req.query;
     if (typeof data.url !== 'string') {
       req.flash('errors', { msg: 'No URL supplied with story' });
-      return next(new TypeError('No URL supplied with story'));
+      return res.redirect('/news');
     }
     var cleanedData = cleanData(data.url);
 
@@ -170,7 +211,6 @@ module.exports = function(app) {
           title: story.headline,
           link: story.link,
           originalStoryLink: dashedName,
-          originalStoryAuthorEmail: story.author.email || '',
           author: story.author,
           rank: story.upVotes.length,
           upVotes: story.upVotes,
@@ -186,45 +226,75 @@ module.exports = function(app) {
     );
   }
 
-  function getStories(req, res, next) {
-    MongoClient.connect(secrets.db, function(err, database) {
-      if (err) {
-        return next(err);
-      }
-      database.collection('stories').find({
-        '$text': {
-          '$search': req.body.data ? req.body.data.searchValue : ''
-        }
-      }, {
-        headline: 1,
-        timePosted: 1,
-        link: 1,
-        description: 1,
-        rank: 1,
-        upVotes: 1,
-        author: 1,
-        image: 1,
-        storyLink: 1,
-        metaDescription: 1,
-        textScore: {
-          $meta: 'textScore'
-        }
-      }, {
-        sort: {
-          textScore: {
-            $meta: 'textScore'
-          }
-        }
-      }).toArray(function(err, items) {
+  function userStories({ body: { search = '' } = {} }, res, next) {
+    if (!search || typeof search !== 'string') {
+      return res.sendStatus(404);
+    }
+
+    return app.dataSources.db.connector
+      .collection('story')
+      .find({
+        'author.username': search.toLowerCase().replace('$', '')
+      })
+      .toArray(function(err, items) {
         if (err) {
           return next(err);
         }
-        if (items !== null && items.length !== 0) {
+        if (items && items.length !== 0) {
+          return res.json(items.sort(sortByRank));
+        }
+        return res.sendStatus(404);
+      });
+  }
+
+  function getStories({ body: { search = '' } = {} }, res, next) {
+    if (!search || typeof search !== 'string') {
+      return res.sendStatus(404);
+    }
+
+    const query = {
+      '$text': {
+        // protect against NoSQL injection
+        '$search': search.replace('$', '')
+      }
+    };
+
+    const fields = {
+      headline: 1,
+      timePosted: 1,
+      link: 1,
+      description: 1,
+      rank: 1,
+      upVotes: 1,
+      author: 1,
+      image: 1,
+      storyLink: 1,
+      metaDescription: 1,
+      textScore: {
+        $meta: 'textScore'
+      }
+    };
+
+    const options = {
+      sort: {
+        textScore: {
+          $meta: 'textScore'
+        }
+      }
+    };
+
+    return app.dataSources.db.connector
+      .collection('story')
+      .find(query, fields, options)
+      .toArray(function(err, items) {
+        if (err) {
+          return next(err);
+        }
+        if (items && items.length !== 0) {
           return res.json(items);
         }
         return res.sendStatus(404);
       });
-    });
   }
 
   function upvote(req, res, next) {
@@ -265,8 +335,11 @@ module.exports = function(app) {
   }
 
   function newStory(req, res, next) {
-    if (!req.user) {
-      return next(new Error('Must be logged in'));
+    if (!req.user.isGithubCool) {
+      req.flash('errors', {
+        msg: 'You must authenticate with Github to post to Camper News'
+      });
+      return res.redirect('/news');
     }
     var url = req.body.data.url;
 
@@ -373,13 +446,11 @@ module.exports = function(app) {
           author: {
             picture: req.user.picture,
             userId: req.user.id,
-            username: req.user.username,
-            email: req.user.email
+            username: req.user.username
           },
           image: data.image,
           storyLink: storyLink,
-          metaDescription: data.storyMetaDescription,
-          originalStoryAuthorEmail: req.user.email
+          metaDescription: data.storyMetaDescription
         });
         return saveInstance(newStory);
       });
