@@ -1,58 +1,49 @@
 import _ from 'lodash';
-import async from 'async';
+import dedent from 'dedent';
 import moment from 'moment';
+import { Observable } from 'rx';
 import debugFactory from 'debug';
 
-import { ifNoUser401 } from '../utils/middleware';
+import {
+  frontEndChallengeId,
+  backEndChallengeId
+} from '../utils/constantStrings.json';
+import { ifNoUser401, ifNoUserRedirectTo } from '../utils/middleware';
+import { observeQuery } from '../utils/rx';
+import { calcCurrentStreak, calcLongestStreak } from '../utils/user-stats';
 
 const debug = debugFactory('freecc:boot:user');
-const daysBetween = 1.5;
+const sendNonUserToMap = ifNoUserRedirectTo('/map');
 
-function calcCurrentStreak(cals) {
-  const revCals = cals.concat([Date.now()]).slice().reverse();
-  let streakBroken = false;
-  const lastDayInStreak = revCals
-    .reduce((current, cal, index) => {
-      const before = revCals[index === 0 ? 0 : index - 1];
-      if (
-        !streakBroken &&
-        moment(before).diff(cal, 'days', true) < daysBetween
-      ) {
-        return index;
-      }
-      streakBroken = true;
-      return current;
-    }, 0);
-
-  const lastTimestamp = revCals[lastDayInStreak];
-  return Math.ceil(moment().diff(lastTimestamp, 'days', true));
+function replaceScriptTags(value) {
+  return value
+    .replace(/<script>/gi, 'fccss')
+    .replace(/<\/script>/gi, 'fcces');
 }
 
-function calcLongestStreak(cals) {
-  let tail = cals[0];
-  const longest = cals.reduce((longest, head, index) => {
-    const last = cals[index === 0 ? 0 : index - 1];
-    // is streak broken
-    if (moment(head).diff(last, 'days', true) > daysBetween) {
-      tail = head;
-    }
-    if (dayDiff(longest) < dayDiff([head, tail])) {
-      return [head, tail];
-    }
-    return longest;
-  }, [cals[0], cals[0]]);
-
-  return Math.ceil(dayDiff(longest));
+function replaceFormAction(value) {
+  return value.replace(/<form[^>]*>/, function(val) {
+    return val.replace(/action(\s*?)=/, 'fccfaa$1=');
+  });
 }
 
-function dayDiff([head, tail]) {
-  return moment(head).diff(tail, 'days', true);
+function encodeFcc(value = '') {
+  return replaceScriptTags(replaceFormAction(value));
 }
 
 module.exports = function(app) {
   var router = app.loopback.Router();
   var User = app.models.User;
-  var Story = app.models.Story;
+  function findUserByUsername$(username, fields) {
+    return observeQuery(
+      User,
+      'findOne',
+      {
+        where: { username },
+        fields
+      }
+    );
+  }
 
   router.get('/login', function(req, res) {
     res.redirect(301, '/signin');
@@ -68,15 +59,40 @@ module.exports = function(app) {
   router.post('/reset-password', postReset);
   router.get('/email-signup', getEmailSignup);
   router.get('/email-signin', getEmailSignin);
-  router.get('/account/api', getAccountAngular);
+  router.get(
+    '/toggle-lockdown-mode',
+    sendNonUserToMap,
+    toggleLockdownMode
+  );
   router.post(
     '/account/delete',
     ifNoUser401,
     postDeleteAccount
   );
-  router.get('/account/unlink/:provider', getOauthUnlink);
-  router.get('/account', getAccount);
-  // Ensure this is the last route!
+  router.get(
+    '/account',
+    sendNonUserToMap,
+    getAccount
+  );
+  router.get('/vote1', vote1);
+  router.get('/vote2', vote2);
+
+  // Ensure these are the last routes!
+  router.get(
+    '/:username/front-end-certification',
+    showCert
+  );
+
+  router.get(
+    '/:username/full-stack-certification',
+    (req, res) => res.redirect(req.url.replace('full-stack', 'back-end'))
+  );
+
+  router.get(
+    '/:username/back-end-certification',
+    showCert
+  );
+
   router.get('/:username', returnUser);
 
   app.use(router);
@@ -86,7 +102,7 @@ module.exports = function(app) {
       return res.redirect('/');
     }
     res.render('account/signin', {
-      title: 'Free Code Camp Login'
+      title: 'Sign in to Free Code Camp using a Social Media Account'
     });
   }
 
@@ -100,7 +116,7 @@ module.exports = function(app) {
       return res.redirect('/');
     }
     res.render('account/email-signin', {
-      title: 'Sign in to your Free Code Camp Account'
+      title: 'Sign in to Free Code Camp using your Email Address'
     });
   }
 
@@ -109,50 +125,36 @@ module.exports = function(app) {
       return res.redirect('/');
     }
     res.render('account/email-signup', {
-      title: 'Create Your Free Code Camp Account'
+      title: 'Sign up for Free Code Camp using your Email Address'
     });
   }
 
   function getAccount(req, res) {
-    if (!req.user) {
-      return res.redirect('/');
-    }
-    res.render('account/account', {
-      title: 'Manage your Free Code Camp Account'
-    });
-  }
-
-  function getAccountAngular(req, res) {
-    res.json({
-      user: req.user || {}
-    });
+    const { username } = req.user;
+    return res.redirect('/' + username);
   }
 
   function returnUser(req, res, next) {
     const username = req.params.username.toLowerCase();
     const { path } = req;
     User.findOne(
-      { where: { username } },
-      function(err, user) {
+      {
+        where: { username },
+        include: 'pledge'
+      },
+      function(err, profileUser) {
         if (err) {
           return next(err);
         }
-        if (!user) {
+        if (!profileUser) {
           req.flash('errors', {
             msg: `404: We couldn't find path ${ path }`
           });
           return res.redirect('/');
         }
-        if (!user.isGithubCool && !user.isMigrationGrandfathered) {
-          req.flash('errors', {
-            msg: `
-              user ${ username } has not completed account signup
-            `
-          });
-          return res.redirect('/');
-        }
+        profileUser = profileUser.toJSON();
 
-        var cals = user
+        var cals = profileUser
           .progressTimestamps
           .map(objOrNum => {
             return typeof objOrNum === 'number' ?
@@ -161,10 +163,10 @@ module.exports = function(app) {
           })
           .sort();
 
-        user.currentStreak = calcCurrentStreak(cals);
-        user.longestStreak = calcLongestStreak(cals);
+        profileUser.currentStreak = calcCurrentStreak(cals);
+        profileUser.longestStreak = calcLongestStreak(cals);
 
-        const data = user
+        const data = profileUser
           .progressTimestamps
           .map((objOrNum) => {
             return typeof objOrNum === 'number' ?
@@ -179,37 +181,182 @@ module.exports = function(app) {
             return data;
           }, {});
 
-        const challenges = user.completedChallenges.filter(function(obj) {
+        const baseAndZip = profileUser.completedChallenges.filter(
+          function(obj) {
           return obj.challengeType === 3 || obj.challengeType === 4;
+          }
+        );
+
+        const bonfires = profileUser.completedChallenges.filter(function(obj) {
+          return (obj.name || '').match(/^Bonfire/g);
         });
 
-        const bonfires = user.completedChallenges.filter(function(obj) {
-          return obj.challengeType === 5 && (obj.name || '').match(/Bonfire/g);
+        const waypoints = profileUser.completedChallenges.filter(function(obj) {
+          return (obj.name || '').match(/^Waypoint/i);
         });
 
         res.render('account/show', {
-          title: 'Camper ' + user.username + '\'s portfolio',
-          username: user.username,
-          name: user.name,
-          isMigrationGrandfathered: user.isMigrationGrandfathered,
-          isGithubCool: user.isGithubCool,
-          location: user.location,
-          github: user.githubURL,
-          linkedin: user.linkedin,
-          google: user.google,
-          facebook: user.facebook,
-          twitter: user.twitter,
-          picture: user.picture,
-          progressTimestamps: user.progressTimestamps,
+          title: 'Camper ' + profileUser.username + '\'s Code Portfolio',
+          username: profileUser.username,
+          name: profileUser.name,
+
+          isMigrationGrandfathered: profileUser.isMigrationGrandfathered,
+          isGithubCool: profileUser.isGithubCool,
+          isLocked: !!profileUser.isLocked,
+
+          pledge: profileUser.pledge,
+
+          isFrontEndCert: profileUser.isFrontEndCert,
+          isBackEndCert: profileUser.isBackEndCert,
+          isFullStackCert: profileUser.isFullStackCert,
+          isHonest: profileUser.isHonest,
+
+          location: profileUser.location,
           calender: data,
-          challenges: challenges,
-          bonfires: bonfires,
-          moment: moment,
-          longestStreak: user.longestStreak,
-          currentStreak: user.currentStreak
+
+          github: profileUser.githubURL,
+          linkedin: profileUser.linkedin,
+          google: profileUser.google,
+          facebook: profileUser.facebook,
+          twitter: profileUser.twitter,
+          picture: profileUser.picture,
+
+          progressTimestamps: profileUser.progressTimestamps,
+
+          baseAndZip,
+          bonfires,
+          waypoints,
+          moment,
+
+          longestStreak: profileUser.longestStreak,
+          currentStreak: profileUser.currentStreak,
+
+          encodeFcc
         });
       }
     );
+  }
+
+  function showCert(req, res, next) {
+    const username = req.params.username.toLowerCase();
+    const { user } = req;
+    const whichCert = req.path.split('/').pop();
+    const showFront = whichCert === 'front-end-certification';
+    Observable.just(user)
+      .flatMap(user => {
+        if (user && user.username === username) {
+          return Observable.just(user);
+        }
+        return findUserByUsername$(username, {
+          isGithubCool: true,
+          isFrontEndCert: true,
+          isFullStackCert: true,
+          isBackEndCert: true,
+          isHonest: true,
+          completedChallenges: true,
+          username: true,
+          name: true
+        });
+      })
+      .subscribe(
+        (user) => {
+          if (!user) {
+            req.flash('errors', {
+              msg: `404: We couldn't find the user ${username}`
+            });
+            return res.redirect('/');
+          }
+          if (!user.isGithubCool) {
+            req.flash('errors', {
+              msg: dedent`
+                This user needs to link GitHub with their account
+                in order to display this certificate to the public.
+              `
+            });
+            return res.redirect('back');
+          }
+          if (user.isLocked) {
+            req.flash('errors', {
+              msg: dedent`
+                ${username} has chosen to hide their work from the public.
+                They need to unhide their work in order for this certificate to
+                be verifiable.
+              `
+            });
+            return res.redirect('back');
+          }
+          if (!user.isHonest) {
+            req.flash('errors', {
+              msg: dedent`
+                ${username} has not agreed to our Academic Honesty Pledge yet.
+              `
+            });
+            return res.redirect('back');
+          }
+
+          if (
+            showFront && user.isFrontEndCert ||
+            !showFront && user.isBackEndCert
+          ) {
+            var { completedDate = new Date() } =
+              _.find(user.completedChallenges, {
+                id: showFront ?
+                  frontEndChallengeId :
+                  backEndChallengeId
+              }) || {};
+
+            return res.render(
+              showFront ?
+                'certificate/front-end.jade' :
+                'certificate/back-end.jade',
+              {
+                username: user.username,
+                date: moment(new Date(completedDate))
+                  .format('MMMM, Do YYYY'),
+                name: user.name
+              }
+            );
+          }
+          req.flash('errors', {
+            msg: showFront ?
+              `Looks like user ${username} is not Front End certified` :
+              `Looks like user ${username} is not Back End certified`
+          });
+          res.redirect('back');
+        },
+        next
+      );
+  }
+
+  function toggleLockdownMode(req, res, next) {
+    if (req.user.isLocked === true) {
+      req.user.isLocked = false;
+      return req.user.save(function(err) {
+        if (err) { return next(err); }
+
+        req.flash('success', {
+          msg: dedent`
+            Other people can now view all your challenge solutions.
+            You can change this back at any time in the "Manage My Account"
+            section at the bottom of this page.
+          `
+        });
+        res.redirect('/' + req.user.username);
+      });
+    }
+    req.user.isLocked = true;
+    return req.user.save(function(err) {
+      if (err) { return next(err); }
+
+      req.flash('success', {
+        msg: dedent`
+          All your challenge solutions are now hidden from other people.
+          You can change this back at any time in the "Manage My Account"
+          section at the bottom of this page.
+        `
+      });
+      res.redirect('/' + req.user.username);
+    });
   }
 
   function postDeleteAccount(req, res, next) {
@@ -221,32 +368,13 @@ module.exports = function(app) {
     });
   }
 
-  function getOauthUnlink(req, res, next) {
-    var provider = req.params.provider;
-    User.findById(req.user.id, function(err, user) {
-      if (err) { return next(err); }
-
-      user[provider] = null;
-      user.tokens =
-        _.reject(user.tokens, function(token) {
-          return token.kind === provider;
-        });
-
-      user.save(function(err) {
-        if (err) { return next(err); }
-        req.flash('info', { msg: provider + ' account has been unlinked.' });
-        res.redirect('/account');
-      });
-    });
-  }
-
   function getReset(req, res) {
     if (!req.accessToken) {
       req.flash('errors', { msg: 'access token invalid' });
       return res.render('account/forgot');
     }
     res.render('account/reset', {
-      title: 'Password Reset',
+      title: 'Reset your Password',
       accessToken: req.accessToken.id
     });
   }
@@ -281,11 +409,6 @@ module.exports = function(app) {
     });
   }
 
-  /**
-  * POST /forgot
-  * Create a random token, then the send user an email with a reset link.
-  */
-
   function postForgot(req, res) {
     const errors = req.validationErrors();
     const email = req.body.email.toLowerCase();
@@ -299,7 +422,7 @@ module.exports = function(app) {
       email: email
     }, function(err) {
       if (err) {
-        req.flash('errors', err);
+        req.flash('errors', err.message);
         return res.redirect('/forgot');
       }
 
@@ -312,6 +435,7 @@ module.exports = function(app) {
     });
   }
 
+  /*
   function updateUserStoryPictures(userId, picture, username, cb) {
     Story.find({ 'author.userId': userId }, function(err, stories) {
       if (err) { return cb(err); }
@@ -331,5 +455,36 @@ module.exports = function(app) {
         cb();
       });
     });
+  }
+  */
+
+  function vote1(req, res, next) {
+    if (req.user) {
+      req.user.tshirtVote = 1;
+      req.user.save(function(err) {
+        if (err) { return next(err); }
+
+        req.flash('success', { msg: 'Thanks for voting!' });
+        res.redirect('/map');
+      });
+    } else {
+      req.flash('error', { msg: 'You must be signed in to vote.' });
+      res.redirect('/map');
+    }
+  }
+
+  function vote2(req, res, next) {
+    if (req.user) {
+      req.user.tshirtVote = 2;
+      req.user.save(function(err) {
+        if (err) { return next(err); }
+
+        req.flash('success', { msg: 'Thanks for voting!' });
+        res.redirect('/map');
+      });
+    } else {
+      req.flash('error', {msg: 'You must be signed in to vote.'});
+      res.redirect('/map');
+    }
   }
 };
