@@ -12,20 +12,27 @@ const pathsOfNoReturn = [
   'css'
 ];
 
+const pathsWhiteList = [
+  'news',
+  'challenges',
+  'map',
+  'news'
+];
+
 const pathsOfNoReturnRegex = new RegExp(pathsOfNoReturn.join('|'), 'i');
+const whiteListRegex = new RegExp(pathsWhiteList.join('|'), 'i');
 
 export default function addReturnToUrl() {
   return function(req, res, next) {
     // Remember original destination before login.
     var path = req.path.split('/')[1];
 
-    if (req.method !== 'GET') {
-      return next();
-    }
-    if (pathsOfNoReturnRegex.test(path)) {
-      return next();
-    }
-    if (/\/stories\/\w+/i.test(req.path)) {
+    if (
+      req.method !== 'GET' ||
+      pathsOfNoReturnRegex.test(path) ||
+      !whiteListRegex.test(path) ||
+      (/news/i).test(path) && (/hot/i).test(req.path)
+    ) {
       return next();
     }
     req.session.returnTo = req.path;
