@@ -1,4 +1,8 @@
+var webpack = require('webpack');
 var path = require('path');
+var webpack = require('webpack');
+
+var __DEV__ = process.env.NODE_ENV !== 'production';
 
 module.exports = {
   entry: './client',
@@ -27,5 +31,14 @@ module.exports = {
       }
     ]
   },
-  plugins: []
+  plugins: [
+    new webpack.optimize.DedupePlugin(),
+    new webpack.optimize.OccurenceOrderPlugin(true),
+    new webpack.DefinePlugin({
+      'process.env': {
+        'NODE_ENV': JSON.stringify(__DEV__ ? 'development' : 'production')
+      },
+      '__DEVTOOLS__': !__DEV__
+    })
+  ]
 };
